Correctly handle argument groups in argparse
Instead of trying to keep track ourselves of the valid argument dests,
peek at the (private!) _actions member. This is relying on internal
details, but in reality the argparser is pretty reliant on the internals
anyway, and this fixes a number of related bugs where more esoteric ways
of accessing arguments weren't being intercepted.

Closes SR-462.